Linear program that selects which controls to use in order to optimize balance

This linear program is used by optimize_controls to choose which controls to select.

balance_LP(
  z,
  X,
  importances,
  st,
  st_vals,
  S,
  q_s,
  N,
  solver,
  integer,
  time_limit
)

Arguments

z
a treatment indicator vector with ith entry equal to 0 if unit i is a control and equal to 1 if unit i is treated.

X
a matrix or data frame containing constraints in the columns. The number of rows should equal the length of z. Balance is achieved when a constraint sums to 0, such that numbers closer to 0 are better. When a constraint does not apply to a particular unit, the entry should be NA. This should typically be generated using generate_constraints.

importances
a vector with length equal to the number of constraints or columns in X. This can be generated using generate_constraints and each nonnegative value denotes how much to prioritize each constraint, with the default being 1 for all constraints.

st
a stratum vector with the ith entry equal to the stratum of unit i. This should have the same order of units and length as z.

st_vals
the unique stratum levels contained in st.

S
the number of unique stratum levels contained in st.

q_s
a named vector indicating how many control units are to be selected from each stratum.

N
the total number of available controls in the data.

solver
a character stating which solver to use to run the linear program. Options are "Rglpk" (default) or "gurobi". You must have gurobi installed to use "gurobi". If gurobi is available, this is the recommended solver.

integer
a logical stating whether to use a mixed integer programming solver instead of randomized rounding. Default is FALSE.

time_limit
numeric stating maximum amount of seconds for which the program is allowed to run before aborting. Default is Inf for no time limit.