Calculate desired number of controls per stratum

Figure out how many units to take from each stratum when some strata are deficient. The result should be used as an input to optimize_controls().

generate_qs(
  z,
  st,
  ratio = NULL,
  max_ratio = NULL,
  max_extra_s = 5,
  strata_dist = NULL
)

Arguments

z
a treatment indicator vector with ith entry equal to 0 if unit i is a control and equal to 1 if unit i is treated.

st
a stratum vector with the ith entry equal to the stratum of unit i. This should have the same order of units and length as z.

ratio
a numeric specifying the desired ratio of controls to treated in each stratum.

max_ratio
a numeric specifying the maximum ratio to allow in a stratum to achieve the overall ratio specified. If NULL, it is set by default to 1.1 times the desired ratio. To have no maximum ratio, set this to Inf.

max_extra_s
single numeric or named vector with values corresponding to the maximum desired number of extra controls to be chosen from each stratum to achieve the overall ratio specified. If this is a vector, the names should correspond to the stratum values from st. The default is 5 for each stratum. To have no maximum, set this to Inf. If both max_ratio and max_s are specified, the maximum of the two will be used for each stratum.

strata_dist
matrix with both row and column names with names corresponding to the stratum values from st and entries corresponding to the distance associated with taking a control from the stratum associated with the row when the desired stratum is the one associated with the column. Lower distance values are more desirable replacements. Typically the diagonal should be 0, meaning there is no penalty for choosing a unit from the correct stratum.

Value

A named vector stating how many controls to take from each stratum.