Calculate standardized differences

Calculate standardized differences in means between treated and control groups, before and after refining the control group. Used within the check_balance function.

get_stand_diffs(
  data,
  z,
  selected,
  st = NULL,
  ist = NULL,
  denom_variance = "treated"
)

Arguments

data
A data frame with columns for which the standardized differences should be calculated.

z
a treatment indicator vector with ith entry equal to 0 if unit i is a control and equal to 1 if unit i is treated.

selected
a boolean vector including whether each unit was selected as part of the treated and control groups for analysis. Should be the same length as z and typically comes from the results of optimize_controls().

st
a stratum vector with the ith entry equal to the stratum of unit i. This should have the same order of units and length as z.

ist
The specific stratum for which the standardized differences should be calculated.

denom_variance
character stating what variance to use in the standardization: either the default "treated", meaning the standardization will use the treated variance (across all strata), or "pooled", meaning the standardization will use the average of the treated and control variances.

Value

data frame containing two columns, one for standardized differences before choosing a subset of controls, and one for after. The rows pertain to covariates.