Standardize covariate vector for balance constraint

This function is used by generate_constraints to standardize covariate vectors to become balance constraints. This standardization is done such that the balance constraint will be minimized when the treated and control groups (within or across strata) have equal means. The function subtracts the treated mean (within or across strata) and divides by the treated or pooled standard deviation (across strata).

stand(z, x, st, ist = NULL, denom_variance = "treated", autogen_missing = 50)

Arguments

z
a treatment indicator vector with ith entry equal to 0 if unit i is a control and equal to 1 if unit i is treated.

x
a covariate vector with ith entry equal to the covariate value of unit i. This should have the same order of units and length as z.

st
a stratum vector with the ith entry equal to the stratum of unit i. This should have the same order of units and length as z.

ist
an optional specification of the target stratum within which the balance constraint is desired. Must be one of the values within st. By default, this is NULL, meaning the generated constraint balances across strata.

denom_variance
character stating what variance to use in the standardization: either the default "treated", meaning the standardization will use the treated variance (across all strata), or "pooled", meaning the standardization will use the average of the treated and control variances.

autogen_missing
whether to automatically generate missingness constraint and how heavily to prioritize it. Should be a numeric or NULL value. NULL indicates that a constraint to balance the rate of missingness (denoted by NA in x) should not be automatically generated. Note that this is not recommended unless the user has already accounted for missing values. If not NULL, autogen_missing should be a numeric stating how heavily to prioritize generated missingness constraints over covariate constraint. The default is 50.

Value

A list with two components:

covariate

a balance constraint for the standardized covariate values of either all treated and control units, or just those in stratum ist.

missingness

a corresponding balance constraint for the rate of missingness if autogen_missing not NULL, otherwise NULL.