Select control units that optimize covariate balance

Select control units within strata that optimize covariate balance. Uses randomized rounding of a linear program or a mixed integer linear program.

optimize_controls(
  z,
  X,
  st,
  importances = NULL,
  ratio = NULL,
  q_s = NULL,
  integer = FALSE,
  solver = "Rglpk",
  seed = NULL,
  runs = 1,
  time_limit = Inf
)

Arguments

z
a treatment indicator vector with ith entry equal to 0 if unit i is a control and equal to 1 if unit i is treated.

X
a matrix or data frame containing constraints in the columns. The number of rows should equal the length of z. Balance is achieved when a constraint sums to 0, such that numbers closer to 0 are better. When a constraint does not apply to a particular unit, the entry should be NA. This should typically be generated using generate_constraints.

st
a stratum vector with the ith entry equal to the stratum of unit i. This should have the same order of units and length as z.

importances
a vector with length equal to the number of constraints or columns in X. This can be generated using generate_constraints and each nonnegative value denotes how much to prioritize each constraint, with the default being 1 for all constraints.

ratio
a numeric specifying the desired ratio of controls to treated in each stratum. If NULL, q_s should be specified.

q_s
a named vector indicating how many control units are to be selected from each stratum. If NULL, ratio should be specified. If both are specified, q_s will take priority. Typically, if the desired ratio is not feasible for every stratum, q_s should be generated using generate_qs.

integer
a logical stating whether to use a mixed integer programming solver instead of randomized rounding. Default is FALSE.

solver
a character stating which solver to use to run the linear program. Options are "Rglpk" (default) or "gurobi". You must have gurobi installed to use "gurobi". If gurobi is available, this is the recommended solver.

seed
the seed to use when doing the randomized rounding of the linear program. This will allow results to be reproduced if desired. The default is NULL, which will choose a random seed to use and return.

runs
the number of times to run randomized rounding of the linear solution. The objective values of all runs will be reported, but the detailed results will only be reported for the run with the lowest objective value. The default is 1.

time_limit
numeric stating maximum amount of seconds for which the program is allowed to run before aborting. Default is Inf for no time limit.

Value

List containing:

objective

objective value of the randomized rounding or mixed integer linear program solution.

objective_wo_importances

objective value of the randomized rounding or mixed integer linear program solution not weighted by the variable importances.

eps

the amount of imbalance obtained in each constraint from the linear program.

importances

the importance of each on the balance constraints.

selected

the selected treated and control units.

controls

the linear program weight assigned to each control and whether it was selected by randomized rounding.

rrdetails

A list containing:

seed

the seed used before commencing the random sampling.

raw_objective

objective value of the randomized rounding or mixed integer linear program solution before the denominator has been corrected for the number of units chosen with missing covariate values.

raw_objective_wo_importances

objective value of the randomized rounding or mixed integer linear program solution not weighted by the variable importances before the denominator has been corrected for the number of units chosen with missing covariate values.

raw_eps

the amount of imbalance obtained in each constraint from the linear program, before the denominators have been corrected for the number of units chosen with missing covariate values.

run_raw_objectives

the objective values for each run of randomized rounding, before denominators have been corrected for missingness.

run_raw_objectives_wo_importances

the objective values for each run of randomized rounding, before denominators have been corrected for missingness, not scaled by constraint importances.

run_objectives

the objective values for each run of randomized rounding.

run_objectives_wo_importances

the objective values for each run of randomized rounding, not scaled by constraint importances.

lpdetails

the full return of the function Rglpk_solve_LP or gurobi plus information about the epsilons and objective values for the linear program solution.

Examples

data('nh0506')

# Create strata
age_cat <- cut(nh0506$age,
               breaks = c(19, 39, 50, 85),
               labels = c('< 40 years', '40 - 50 years', '> 50 years'))
strata <- age_cat : nh0506$sex

# Balance age, race, education, poverty ratio, and bmi both across and within the levels of strata
constraints <- generate_constraints(
                 balance_formulas = list(age + race + education + povertyr + bmi ~ 1 + strata),
                 z = nh0506$z,
                 data = nh0506)

#> Error in eval(predvars, data, env): object 'strata' not found

# Choose one control for every treated unit in each stratum,
# balancing the covariates as described by the constraints
results <- optimize_controls(z = nh0506$z,
                             X = constraints$X,
                             st = strata,
                             importances = constraints$importances,
                             ratio = 1)

#> Error in is.data.frame(X): object 'constraints' not found

# If you want to use a ratio that's not feasible,
# you can supply a vector of the desired number of controls per stratum, q_s,
# typically generated by creating a distance matrix between strata and using
# generate_qs():

age_dist <- matrix(data = c(0, 1, 2, 1, 0, 1, 2, 1, 0),
                   nrow = 3,
                   byrow = TRUE,
                   dimnames = list(levels(age_cat), levels(age_cat)))

sex_dist <- matrix(data = c(0, 1, 1, 0),
                   nrow = 2,
                   dimnames = list(levels(nh0506$sex), levels(nh0506$sex)))

strata_dist <- create_dist_matrix(age_dist, sex_dist)

qs <- generate_qs(z = nh0506$z,
                  st = strata,
                  ratio = 2.5,
                  max_ratio = 2.6,
                  max_extra_s = 0,
                  strata_dist = strata_dist)

results <- optimize_controls(z = nh0506$z,
                             X = constraints$X,
                             st = strata,
                             importances = constraints$importances,
                             q_s = qs)

#> Error in is.data.frame(X): object 'constraints' not found